Finish the encoding part of SatSolver
It's now capable of solving the 11th problem: it prints out DIMACS task
that I can paste into Minisat, which spews out this:

    1 -2 -3 -4 5 -6 -7 -8 9 0

Which, in English, means that variables 1, 5, and 9 are set to true.
Variables get their numbers according to the formula

    1 + f*H + h

where f is the figure's vertex index (counting from 0)
      h is the hole's vertex index (counting from 0)
      H is the number of vertices that comprise the hole

The variable being true means that f-vertex maps to h-vertex. So the
above result means that:

- 0th vertex of the figure maps to the 0th vertex of the hole
- 1st -> 1st
- 2nd -> 2nd

Which is indeed a valid solution to the problem.